backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.1
dotenv==0.9.9
exceptiongroup==1.3.0
fastapi==0.121.2
h11==0.16.0
idna==3.11
numpy==2.2.6
pandas==2.3.3
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
